requirements.txt
django==4.2.16
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
pytest==8.3.3
django-cors-headers==4.4.0
drf-spectacular==0.27.2
pytest-django==4.9.0
psycopg2-binary==2.9.9
pytest-factoryboy==2.7.0
django-unfold==0.40.0
django-redis==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1